pyproject.toml
[tool.poetry]
name = "flask-api-template"
version = "0.1.0"
description = ""
authors = ["Stephen Chung <chung25stephen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
werkzeug = "^2.2.3"
flask-restx = "^1.1.0"
mysql-connector-python = "^8.0.32"
psycopg2 = "^2.9.5"
flask-sqlalchemy = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
